requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# gpt-chatbot-client.py: 1,5
# gptchatbot.py: 7,9,10
discord.py == 1.0.0a1765+g5e65ec9

# gptchatbot.py: 13
# src/chatbot.py: 4
# src/model.py: 1
numpy == 1.14.2

# src/encoder.py: 5
regex == 2017.4.5

# gptchatbot.py: 14
# src/chatbot.py: 5
# src/model.py: 2,3
# src/sample.py: 1
tensorflow == 1.12.0
